Revert "msm: camera: Remove function call in macro expansion"
This reverts commit a743a4c087fc3e42d42a3ad49cc7af9d4c935492.

+++ drivers/media/platform/msm/camera/cam_cpas/cam_cpas_hw.c
@@ -585,7 +585,7 @@ static int cam_cpas_util_apply_client_axi_vote(
 		if (rc) {
 			CAM_ERR(CAM_CPAS,
 				"Failed camnoc vote ab[%llu] ib[%llu] rc=%d",
-				(uint64_t)0, camnoc_bw, rc);
+				0, camnoc_bw, rc);
 			goto unlock_axi_port;
 		}
 	}
